add index.js
