+++ modules/rubiks.js
@@ -1,6 +1,5 @@
 import { glMatrix, mat4, vec3, mat3 } from './modules/gl-matrix-min.js';
-import { BoundingBox } from './modules/bouding-box.js';
-import { $ } from 'jquery';
+import BoundingBox from './modules/bouding-box.js';
 
 const EYE = [0, 0, 20];
 const CENTER = [0, 0, 0];
@@ -23,11 +22,6 @@ const LEFT_MOUSE = 0;
 const RIGHT_MOUSE = 2;
 const DEBOUNCE_TIMEOUT = 50;
 
-var $canvas;
-var gl;
-var rubiksCube;
-var shaderProgram;
-
 var leftMouseDown = false;
 var rightMouseDown = false;
 
@@ -35,8 +29,7 @@ var modelViewMatrix = mat4.create();
 var projectionMatrix = mat4.create();
 var rotationMatrix = mat4.create();
 
-function RubiksCube(data) {
-    this.data = data;
+function RubiksCube(data, gl, shaderProgram, $canvas) {
     this.buffers = null;
     this.rotation = null;
     this.scrambleCycles = 0;
@@ -54,7 +47,7 @@ function RubiksCube(data) {
                 for (let b = 0; b < 3; b++) {
                     // Each cube has dimensions 2x2x2 units.
                     const coordinates = vec3.fromValues(2 * (r - 1), 2 * (g - 1), 2 * (b - 1));
-                    const cube = new Cube(this, coordinates, data);
+                    const cube = new Cube(this, coordinates, data, gl, shaderProgram);
                     this.cubes[r][g][b] = cube;
                 }
             }
@@ -64,11 +57,11 @@ function RubiksCube(data) {
     this.initBuffers = function() {
         const vertices = gl.createBuffer();
         gl.bindBuffer(gl.ARRAY_BUFFER, vertices);
-        gl.bufferData(gl.ARRAY_BUFFER, new Float32Array(this.data.vertices), gl.STATIC_DRAW);
+        gl.bufferData(gl.ARRAY_BUFFER, new Float32Array(data.vertices), gl.STATIC_DRAW);
 
         const normals = gl.createBuffer();
         gl.bindBuffer(gl.ARRAY_BUFFER, normals);
-        gl.bufferData(gl.ARRAY_BUFFER, new Float32Array(this.data.normals), gl.STATIC_DRAW);
+        gl.bufferData(gl.ARRAY_BUFFER, new Float32Array(data.normals), gl.STATIC_DRAW);
 
         const buffer = new Array();
         for (let faceGroup of data.faces) {
@@ -338,9 +331,7 @@ function RubiksCube(data) {
     }
 }
 
-function Cube(rubiksCube, coordinates, data) {
-    this.rubiksCube = rubiksCube;
-    this.data = data;
+function Cube(rubiksCube, coordinates, data, gl, shaderProgram) {
     this.rotationMatrix = mat4.create();
     this.coordinates = coordinates;
 
@@ -358,10 +349,10 @@ function Cube(rubiksCube, coordinates, data) {
         const mvMatrix = mat4.create();
         mat4.copy(mvMatrix, modelViewMatrix);
         this.transform();
-        setMatrixUniforms();
+        setMatrixUniforms(gl, shaderProgram);
 
         let offset = 0;
-        for (let faceGroup of this.data.faces) {
+        for (let faceGroup of data.faces) {
             const material = faceGroup.material;
             // Blender doesn't seem to support per-object ambient colors or export the global ambient color,
             // so we compute our own ambient color as a darker version of the diffuse color.
@@ -406,129 +397,7 @@ function debounce(f, timeout) {
     }
 }
 
-function initWebGL(canvas) {
-    if (!window.WebGLRenderingContext) {
-        console.log("Your browser doesn't support WebGL.")
-        return null;
-    }
-    gl = canvas.getContext('webgl', {preserveDrawingBuffer: true}) || canvas.getContext('experimental-webgl', {preserveDrawingBuffer: true});
-    canvas.width = canvas.clientWidth;
-    canvas.height = canvas.clientHeight;
-    window.onresize = function () {
-        canvas.width = canvas.clientWidth;
-        canvas.height = canvas.clientHeight;
-    };
-
-    if (!gl) {
-        console.log("Your browser supports WebGL, but initialization failed.");
-        return null;
-    }
-    return gl;
-}
-
-function getShader(gl, id) {
-    const shaderScript = document.getElementById(id);
-    if (!shaderScript) {
-        return null;
-    }
-    let source = '';
-    let currentChild = shaderScript.firstChild;
-    while (currentChild) {
-        if (currentChild.nodeType === currentChild.TEXT_NODE) {
-            source += currentChild.textContent;
-        }
-        currentChild = currentChild.nextSibling;
-    }
-    let shader;
-    if (shaderScript.type === 'x-shader/x-fragment') {
-        shader = gl.createShader(gl.FRAGMENT_SHADER);
-    } else if (shaderScript.type === 'x-shader/x-vertex') {
-        shader = gl.createShader(gl.VERTEX_SHADER);
-    } else {
-        return null;
-    }
-    gl.shaderSource(shader, source);
-    gl.compileShader(shader);
-    if (!gl.getShaderParameter(shader, gl.COMPILE_STATUS)) {
-        console.log('An error occurred while compiling the shader: ' + gl.getShaderInfoLog(shader));
-        return null;
-    }
-    return shader;
-}
-
-function initShaders() {
-    const fragmentShader = getShader(gl, 'fragmentShader');
-    const vertexShader = getShader(gl, 'vertexShader');
-    shaderProgram = gl.createProgram();
-    gl.attachShader(shaderProgram, fragmentShader);
-    gl.attachShader(shaderProgram, vertexShader);
-    gl.linkProgram(shaderProgram);
-    if (!gl.getProgramParameter(shaderProgram, gl.LINK_STATUS)) {
-        console.log('Unable to initialize the shader program');
-    }
-    gl.useProgram(shaderProgram);
-
-    shaderProgram.vertexPosition = gl.getAttribLocation(shaderProgram, 'vertexPosition');
-    gl.enableVertexAttribArray(shaderProgram.vertexPosition);
-
-    shaderProgram.vertexNormal = gl.getAttribLocation(shaderProgram, 'vertexNormal');
-    gl.enableVertexAttribArray(shaderProgram.vertexNormal);
-
-    shaderProgram.eye = gl.getUniformLocation(shaderProgram, 'eye');
-    gl.uniform3fv(shaderProgram.eye, EYE);
-
-    for (let i = 0; i < LIGHTS.length; i++) {
-        const lightPosition = `lights[${i}].position`;
-        const lightIntensity = `lights[${i}].intensity`;
-        shaderProgram[lightPosition] = gl.getUniformLocation(shaderProgram, lightPosition);
-        shaderProgram[lightIntensity] = gl.getUniformLocation(shaderProgram, lightIntensity);
-        gl.uniform3fv(shaderProgram[lightPosition], LIGHTS[i].position);
-        gl.uniform1f(shaderProgram[lightIntensity], LIGHTS[i].intensity);
-    }
-
-    shaderProgram.lighting = gl.getUniformLocation(shaderProgram, 'lighting');
-    shaderProgram.ambient = gl.getUniformLocation(shaderProgram, 'ambient');
-    shaderProgram.diffuse = gl.getUniformLocation(shaderProgram, 'diffuse');
-    shaderProgram.specular = gl.getUniformLocation(shaderProgram, 'specular');
-    shaderProgram.specularExponent = gl.getUniformLocation(shaderProgram, 'specularExponent');
-}
-
-// timestamp is a DOMHighResTimeStamp.
-// See https://developer.mozilla.org/en-US/docs/Web/API/window/requestAnimationFrame.
-function drawScene() {
-    let startTime = 0;
-
-    const animate = function(timestamp) {
-        const timeDelta = timestamp - startTime;
-
-        rubiksCube.draw();
-        rubiksCube.rotate(timeDelta);
-        requestAnimationFrame(animate);
-
-        startTime = timestamp;
-    }
-    return animate;
-}
-
-function start(data) {
-    gl = initWebGL($canvas[0]);
-    if (gl) {
-        gl.clearColor(1.0, 1.0, 1.0, 1.0);
-        gl.enable(gl.DEPTH_TEST);
-        gl.depthFunc(gl.LEQUAL);
-        gl.enable(gl.CULL_FACE);
-        gl.cullFace(gl.BACK);
-        gl.clear(gl.COLOR_BUFFER_BIT | gl.DEPTH_BUFFER_BIT);
-
-        initShaders();
-
-        rubiksCube = new RubiksCube(data);
-        perspectiveView();
-        drawScene()(performance.now());
-    }
-}
-
-function setMatrixUniforms() {
+function setMatrixUniforms(gl, shaderProgram) {
     const projectionUniform = gl.getUniformLocation(shaderProgram, 'projectionMatrix');
     gl.uniformMatrix4fv(projectionUniform, false, projectionMatrix);
 
@@ -614,27 +483,11 @@ function togglePerspective(event) {
     }
 }
 
-function scramble() {
+function scramble(rubiksCube) {
     if (rubiksCube.scrambleCycles === 0) {
         rubiksCube.scrambleCycles = Math.ceil(Math.random() * 10 + 10); // an integer between 10 and 20
         rubiksCube.scramble();
     }
 }
 
-$(document).ready(function() {
-    $canvas = $('#glcanvas');
-
-    const pathname = location.pathname;
-    const base = pathname.substring(0, pathname.lastIndexOf('/'));
-
-    $.get(`${base}/models/rubiks-cube.json`, function(data) {
-        start(data[0]);
-        $canvas.bind('contextmenu', function() { return false; });
-        $canvas.mousedown(rubiksCube.startRotate.bind(rubiksCube));
-        $canvas.mouseup(rubiksCube.endRotate.bind(rubiksCube));
-        $canvas.mouseout(rubiksCube.endRotate.bind(rubiksCube));
-        $('body').keypress(togglePerspective);
-    });
-});
-
-export { RubiksCube, scramble };
+export { EYE, LIGHTS, RubiksCube, scramble, togglePerspective };
